+++ packages/app-staking/src/Overview/Address.tsx
@@ -42,7 +42,7 @@ interface StakingState {
   isNominatorMe: boolean;
   nominators: [AccountId, Balance][];
   sessionId?: string;
-  stashId?: string;
+  stashId: string;
 }
 
 const WITH_VALIDATOR_PREFS = { validatorPayment: true };
@@ -57,7 +57,8 @@ function Address ({ address, authorsMap, className, defaultName, filter, hasQuer
     balanceOpts: { bonded: true },
     hasNominators: false,
     isNominatorMe: false,
-    nominators: []
+    nominators: [],
+    stashId: address.toString()
   });
 
   useEffect((): void => {
@@ -89,17 +90,15 @@ function Address ({ address, authorsMap, className, defaultName, filter, hasQuer
         ),
         nominators,
         sessionId: nextSessionIds && nextSessionIds[0]?.toString(),
-        stashId: stashId?.toString()
+        stashId: (stashId || address).toString()
       });
     }
   }, [stakingInfo]);
 
   useEffect((): void => {
-    setHasActivity(
-      recentlyOnline && recentlyOnline[stashId || '']
-        ? recentlyOnline[stashId || ''].isOnline
-        : true
-    );
+    if (recentlyOnline && stashId && recentlyOnline[stashId]) {
+      setHasActivity(recentlyOnline[stashId].isOnline);
+    }
   }, [recentlyOnline, stashId]);
 
   if ((filter === 'hasNominators' && !hasNominators) ||
@@ -109,18 +108,6 @@ function Address ({ address, authorsMap, className, defaultName, filter, hasQuer
     (filter === 'iNominated' && !isNominatorMe) ||
     (filter === 'nextSet' && !isElected)) {
     return null;
-  }
-
-  if (!stashId) {
-    return (
-      <AddressCard
-        className={className}
-        defaultName={defaultName}
-        isDisabled
-        value={address}
-        withBalance={false}
-      />
-    );
   }
 
   const lastBlockNumber = authorsMap[stashId];
@@ -188,7 +175,6 @@ function Address ({ address, authorsMap, className, defaultName, filter, hasQuer
           )}
         </>
       }
-      isDisabled={isSubstrateV2 && !hasActivity}
       overlay={
         hasQueries && api.query.imOnline?.authoredBlocks && (
           <Icon
@@ -278,7 +264,6 @@ export default translate(
     }
 
     .staking--Address-info {
-      /* Small additional margin to take care of validator highlights */
       margin-right: 0.25rem;
       text-align: right;