+++ packages/react-api/src/typeRegistry.ts
@@ -1,0 +1,9 @@
+// Copyright 2017-2019 @polkadot/react-api authors & contributors
+// This software may be modified and distributed under the terms
+// of the Apache-2.0 license. See the LICENSE file for details.
+
+import { TypeRegistry } from '@polkadot/types';
+
+const registry = new TypeRegistry();
+
+export default registry;