+++ packages/react-params/src/Param/Struct.tsx
@@ -6,21 +6,27 @@ import { TypeDef } from '@polkadot/types/types';
 import { ParamDef, Props, RawParam } from '../types';
 
 import React, { useEffect, useState } from 'react';
+import { registry } from '@polkadot/react-api';
+import { createType, getTypeDef } from '@polkadot/types';
 
 import Params from '../';
 import Base from './Base';
 import Static from './Static';
-
-import {createType, getTypeDef } from '@polkadot/types';
 
 export default function StructParam (props: Props): React.ReactElement<Props> {
   const { className, isDisabled, label, onChange, style, type, withLabel } = props;
   const [params, setParams] = useState<ParamDef[]>([]);
 
   useEffect((): void => {
-    const rawType = createType(type.type as any).toRawType();
-    const typeDef = getTypeDef(rawType);
-    setParams((typeDef.sub as TypeDef[]).map((subType): ParamDef => ({ name: subType.name, type: subType })));
+    let typeDef;
+    try {
+      const rawType = createType(registry, type.type as any).toRawType();
+      typeDef = getTypeDef(rawType);
+    } catch (e) {
+      typeDef = type;
+    }
+
+    setParams((typeDef.sub as TypeDef[]).map((type): ParamDef => ({ name: type.name, type })));
   }, [type]);
 
   if (isDisabled) {